% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_matrix.R
\name{generate_matrix}
\alias{generate_matrix}
\title{Generate a Random Correlation Matrix}
\usage{
generate_matrix(var_names, range = c(-0.01, 0.01))
}
\arguments{
\item{var_names}{A character vector containing the names of the variables.}

\item{range}{A numeric vector of length 2 specifying the range for the random correlations, default is c(-0.01, 0.01).
The provided values should lie between -1 and 1.}
}
\value{
A symmetric matrix of size length(var_names) x length(var_names) with random correlation values.
}
\description{
This function creates a symmetric matrix of random correlations given a set of variable names.
The diagonal elements are always set to 1 (since any variable has a perfect correlation with itself).
}
\examples{
generate_cor_matrix(c("var1", "var2", "var3"))
generate_cor_matrix(c("var1", "var2"), range = c(-0.5, 0.5))
}
